requirements.txt
# Core dependencies
chromadb==0.3.29
openai==0.28.1
python-dotenv==1.0.1

# ChromaDB dependencies
duckdb==0.9.2
onnxruntime==1.16.3
tokenizers==0.15.0
sentence-transformers==2.2.2
hnswlib>=0.7.0

# Utility dependencies
numpy==1.24.4
pandas==2.1.4
tqdm==4.66.1

# FastAPI dependencies
fastapi==0.85.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
